pyproject.toml
[tool.poetry]
name = "eip1559"
version = "0.1.0"
description = ""
authors = ["Georgios Konstantopoulos <me@gakonst.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.0.4"
matplotlib = "^3.2.1"
numpy = "^1.18.4"
google-cloud-bigquery = {extras = ["bqstorage", "pandas"], version = "^1.25.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
